auth/requirements.txt
fastapi==0.88.0
importlib-metadata==4.2.0
passlib==1.7.4
python-jose==3.3.0
pytz==2022.6
sqlalchemy==1.1.2
uvicorn==0.12.0